fin iteration 1 of code-cleanup
